requirements.txt
flask>=3.0
opencv-python>=4.8
Pillow>=10.0
numpy>=1.24
optimum[openvino]>=1.18
transformers>=4.45
qwen-vl-utils>=0.0.8
openvino>=2024.0
